+++ novae/model.py
@@ -364,7 +364,7 @@ class Novae(L.LightningModule, PyTorchModelHubMixin):
         Args:
             adata: An `AnnData` object, or a list of `AnnData` objects. Optional if the model was initialized with `adata`.
             zero_shot: If `True`, the model will be used in zero-shot mode, i.e. without training. In this case, the model will assign each cell to a leaf based on the latent representations.
-            reference: Use only if `zero_shot=True`. Reference slide to use for the new prototypes. Can be the AnnData index, a unique slide id, or one of `["all", "largest"]`.
+            reference: Use only if `zero_shot=True`. Reference slide to use for the new prototypes. Can be one or multiple AnnData indices, slide ids, or one of `["all", "largest"]`.
             accelerator: Accelerator to use. For instance, `'cuda'`, `'cpu'`, or `'auto'`. See Pytorch Lightning for more details.
             num_workers: Number of workers for the dataloader.
         """
@@ -596,7 +596,7 @@ class Novae(L.LightningModule, PyTorchModelHubMixin):
         self,
         adata: AnnData | list[AnnData],
         *,
-        reference: Literal["all", "largest"] | str | int | list[str] | list[int] = "all",
+        reference: Literal["all", "largest"] | str | int | list[str] | list[int] | None = "all",
         max_epochs: int = 20,
         accelerator: str = "cpu",
         num_workers: int | None = None,
@@ -609,7 +609,7 @@ class Novae(L.LightningModule, PyTorchModelHubMixin):
 
         Args:
             adata: An `AnnData` object, or a list of `AnnData` objects. Optional if the model was initialized with `adata`.
-            reference: Reference slide to use for the new prototypes. Can be the AnnData index, a unique slide id, or one of `["all", "largest"]`.
+            reference: Reference slide to use for the new prototypes. Can be one or multiple AnnData indices, slide ids, or one of `["all", "largest"]`. If `None`, initializes prototypes randomly.
             max_epochs: Maximum number of training epochs.
             accelerator: Accelerator to use. For instance, `'cuda'`, `'cpu'`, or `'auto'`. See Pytorch Lightning for more details.
             num_workers: Number of workers for the dataloader.
@@ -623,7 +623,11 @@ class Novae(L.LightningModule, PyTorchModelHubMixin):
 
         assert adata is not None, "Please provide an AnnData object to fine-tune the model."
 
-        self.init_prototypes(adata, reference=reference)
+        if reference is None:
+            self.swav_head.initialize_prototypes()
+        else:
+            self.init_prototypes(adata, reference=reference)
+
         self.init_slide_queue(adata, min_prototypes_ratio)
 
         self.fit(