+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Full dataset added to [our Hugging Face Hub](https://huggingface.co/datasets/MICS-Lab/novae)
+- Official Docker images released - [see here](https://hub.docker.com/r/quentinblampey/novae)
 
 ### Changed
 - Update hf-hub to use `xet` for faster dataset download from Hugging Face Hub