+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Support `reference=None` in `fine_tune` (random initialization) and used as default
 - Added `fast` argument in `novae.plot.domains` to have a quick (but less accurate) rendering of domains.
 
+### Breaking changes
+- Remove support for `python==3.10`
+
 ### Changed
 - Update hf-hub to use `xet` for faster dataset download from Hugging Face Hub
 - Update to `spatialdata>=0.5.0` to avoid installation issue related to `xarray-dataclasses`